Trying to define some classes and organizing these classes to files
