+++ tools/digital_matcher/main.cpp
@@ -1,409 +1,22 @@
 #include <iostream>
 #include <fstream>
-#include <libfprint/fprint.h>
-#include <openssl/bio.h>
-#include <openssl/evp.h>
-#include <openssl/buffer.h>
-#include <curl/curl.h>
+
 #include <list>
 #include <string.h>
 
-#include <wiringPi.h>
+#if DEBUG == 0
 #define DOOR 15
+#include <wiringPi.h>
+#endif
 
-#include "json.hpp"
+#include "Digital.h"
+#include "lib/json.hpp"
 
 using namespace std;
 using json = nlohmann::json;
 
-/**
- * Shows a fatal error message, unloads fprintlib and exit the program with
- * the status code EXIT_FAILURE
- * @param msg Message to be displayed
- */
-void inline fatalError(string msg){
-    cerr << "❮ ☠ ❯ " << msg << endl;
-    fp_exit();
-    exit(EXIT_FAILURE);
-}
 
-    /**
-     * initialize fprintlib and open a device for use
-     * @return device to be used
-     */
-    inline struct fp_dev* init_libfp(){
-      /**
-       * initialize fprintlib
-       */
-      cout << "❮ ▶ ❯ Initialiaing fingeprint reader..." << endl;
 
-
-      // if the lib couldn't be initialized
-      if (fp_init() < 0) {
-        fatalError("Failed to initialize libfprint!");
-      }
-
-
-      /**
-       * find and open fingerprint reader
-       */
-      struct fp_dscv_dev **devices_found = fp_discover_devs();
-
-      // if no device found, exit the program
-      if (devices_found == NULL || devices_found[0] == NULL) {
-        fatalError("No reader found!");
-      }
-
-      // list devices found
-      cout << "Chosen reader: " << endl;
-      for (int i = 0; devices_found[i] != NULL; i++) {
-        struct fp_dscv_dev *dscv_device = devices_found[i];
-        struct fp_driver *driver = fp_dscv_dev_get_driver(dscv_device);
-        if (i == 0) {
-          cout << "\t➜ ";
-        } else {
-          cout << "\t● ";
-        }
-        cout << fp_driver_get_full_name(driver) << endl;
-      }
-
-      // choose one reader to use, and open it
-      struct fp_dev* device = fp_dev_open(devices_found[0]);
-
-      // check if the reader could be opened
-      if (device == NULL) {
-        fatalError("Couldn't open selected reader for use!");
-      }
-
-      // free list of devices from memory
-      fp_dscv_devs_free(devices_found);
-
-
-      cout << "❮ ✔ ❯ Reader ready for use" << endl << endl;
-      return device;
-    }
-
-
-  public:
-
-    /**
-     * Read cached fingerprints from a file an return it via parameters
-     * @param cacheFileName file to read from
-     * @param cache prints from the file
-     * @param ids user IDs referring to the prints
-     */
-    inline void load_cache(string cacheFileName){
-      cout << "❮ ▶ ❯ Loading fingerprints... " << endl;
-
-      // open cache file
-      ifstream cacheFile("cache");
-
-      // if failed to open file, quit program
-      if (!cacheFile.is_open()) {
-        fatalError("failed to open cache file");
-      }
-
-      // read entries
-      string line;
-      std::list<int> idList;
-      std::list<struct fp_print_data*> printsList;
-      while (getline(cacheFile, line)) {
-        string idString = "";
-        int dataStartPos = -1;
-        for(int i = 0; i < line.size(); i++){
-          // found separator
-          if (line[i] == '|'){
-            dataStartPos = i + 1;
-            break;
-          }
-
-          // add another digiti to the id
-          idString += line[i];
-        }
-
-        /**
-         * add id to list
-         */
-        idList.push_back(std::stoi(idString));
-
-        // cleaning buffer from c???
-        string test = line.substr(dataStartPos);
-        char * test2 = (char*)test.c_str();
-        /**
-         * convert print data from base64 to binary
-         */
-        BIO *bio, *b64;
-        char* b64message = test2;
-        //        b64message[264] = '\0';
-        //        cout << "❮ ☝ ❯ base64: "<< (string(b64message, strlen(b64message))) << endl;
-
-
-        // get original size
-        size_t y = line.size();
-        size_t b64messageSize = strlen(b64message);
-        size_t decodeLen = (3*strlen(b64message))/4;
-
-        // remove padding from size
-        if (line[line.size() - 1] == '='){
-          decodeLen--;
-        }
-        if (line[line.size() - 2] == '='){
-          decodeLen--;
-        }
-
-        unsigned char* buffer = (unsigned char*)malloc(decodeLen + 1);
-
-
-        bio = BIO_new_mem_buf(b64message, -1);
-        b64 = BIO_new(BIO_f_base64());
-        bio = BIO_push(b64, bio);
-
-        BIO_set_flags(bio, BIO_FLAGS_BASE64_NO_NL); //Do not use newlines to flush buffer
-        size_t length = BIO_read(bio, buffer, b64messageSize);
-        BIO_free_all(bio);
-
-        /**
-         * add print data to the list
-         */
-
-        struct fp_print_data* tmp = fp_print_data_from_data(buffer, decodeLen);
-
-        // failed to convert data
-        if (tmp == NULL){
-          fatalError("could not read entry from cache file");
-        }
-
-        // add print to list
-        printsList.push_back(tmp);
-      }
-
-      // close file
-      cacheFile.close();
-
-      /**
-       * convert lists into array and return it
-       */
-      // generate array of ids
-      this->ids = new int[idList.size()];
-      copy(idList.begin(), idList.end(), this->ids);
-
-      // generate array of prints
-      this->cache = new struct fp_print_data*[printsList.size() + 1];
-      (this->cache)[printsList.size()] = NULL;
-      copy(printsList.begin(), printsList.end(), this->cache);
-
-      cout << "❮ ✔ ❯ Fingerprints loaded" << endl << endl;
-    }
-
-
-    Device(){
-      this->device = init_libfp();
-      this->cache = NULL;
-      this->ids = NULL;
-    }
-
-    int scan(){
-      while(true){
-
-        cout << endl << "❮ ☝ ❯ Waiting for finger..." << endl;
-
-        size_t cacheMatchPos = 0;
-        int resultCode = fp_identify_finger(this->device, this->cache, &cacheMatchPos);
-
-        // if there was a error, retry
-        if (resultCode < 0){
-          cerr << "❮ ⚠ ❯ Error matching fingerprint!" << endl;
-          continue;
-        }
-
-        switch (resultCode){
-
-          case FP_VERIFY_NO_MATCH:
-            cout << "❮ ☝ ✖ ❯ Fingerprint does not match any database entry" << endl;
-            break;
-
-          case FP_VERIFY_RETRY:
-          case FP_VERIFY_RETRY_TOO_SHORT:
-          case FP_VERIFY_RETRY_CENTER_FINGER:
-          case FP_VERIFY_RETRY_REMOVE_FINGER:
-            cout << "❮ ☝ ↻ ❯ Failed to read fingerprint, retrying..." << endl;
-            break;
-
-          case FP_VERIFY_MATCH:
-            cout << "❮ ☝ ✔ ❯ Fingerprint match user ID: " << ids[cacheMatchPos] << endl;
-            // open the door
-            digitalWrite(DOOR, HIGH);
-            delay(500);
-            digitalWrite(DOOR, LOW);
-
-            // log on server that the door was opened
-            cout << "❮ ⬆ ❯ uploading log..." << endl;
-            CURL* curl;
-            CURLcode res;
-            curl = curl_easy_init();
-            if (curl == NULL){
-              cerr << "❮ ⚠ ❯ Couldn't get a curl handler!" << endl;
-            }else {
-
-              /* First set the URL that is about to receive our POST. This URL can
-                 just as well be a https:// URL if that is what should receive the
-                 data. */
-              string body = "laboratory_id=2&user_id=" + std::to_string(ids[cacheMatchPos]);
-              //                    cout << body << endl;
-
-              curl_easy_setopt(curl, CURLOPT_URL,
-                  "https://siglaufscar.herokuapp.com/fingerprint/access/");
-              /* Now specify the POST data */
-              curl_easy_setopt(curl, CURLOPT_POSTFIELDS, body.c_str());
-              curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
-
-              /* Perform the request, res will get the return code */
-              res = curl_easy_perform(curl);
-              /* Check for errors */
-              if (res != CURLE_OK) {
-                cerr << "❮ ⚠ ❯ Could not save log on the server! (" << curl_easy_strerror(res) << ")" << endl;
-              }else{
-                cout << "❮ ✔ ❯ Log saved to the server" << endl;
-              }
-
-              /* always cleanup */
-              curl_easy_cleanup(curl);
-            }
-            break;
-        }
-      }
-    }
-
-};
-
-class Digital{
-  private:
-    Device * device;
-    string json_string;
-    json data;
-
-    /*
-     * Callback using curl, when curl download fingerprint's data,
-     * this function converts every string into a json data
-     */
-    size_t write_json(char *ptr, size_t size, size_t nmemb){
-      string aux = (string)ptr;
-      if(aux.size() != nmemb){
-        // it is the last packet, so, we need to remove some chunk bytes there
-        this->json_string.append(ptr,0,((string)ptr).size()-7);
-      }else{
-        this->json_string.append(ptr);
-      }
-      return size*nmemb;
-    }
-
-    /*
-     * A way to use this lib (made in C) to C++
-     */
-
-    static size_t write_callback(char *ptr, size_t size, size_t nmemb, void *userdata){
-      return static_cast<Digital*>(userdata)->write_json(ptr, size, nmemb);
-    }
-
-  public:
-    int init(){
-      /**
-       * initialize wiringPi
-       */
-      cout << "❮ ▶ ❯ Initialiaing wiringPi... " << endl;
-      wiringPiSetup();
-      pinMode(DOOR, OUTPUT);
-      digitalWrite(DOOR, LOW);
-
-      cout << "❮ ✔ ❯ WiringPi ready" << endl << endl;
-
-      /**
-       * initialize libcurl
-       */
-      cout << "❮ ▶ ❯ Initialiaing libcurl... " << endl;
-      curl_global_init(CURL_GLOBAL_ALL);
-      cout << "❮ ✔ ❯ Libcurl ready for use" << endl << endl;
-
-      /*
-       * Download fingerprint's cache
-       */
-      if(this->get_data()){
-        this->write_data("cache");
-      }
-
-      device = new Device();
-
-      device->load_cache("cache");
-      device->scan();
-    }
-
-    /*
-     * Write all fingerprint's data to a file
-     * @param name of the file
-     */
-    void write_data(string name){
-      cout << "❮ ▶ ❯ Writing all fingerprints to cache... " << endl;
-      fstream fs;
-
-      fs.open(name, ios::out|ios::trunc);
-
-      for(json& unique : this->data){
-        string biometric = unique["biometric"];
-        replace(biometric.begin(), biometric.end(), ' ', '+');
-        fs << unique["user_id"] << "|" << biometric << endl;
-      }
-
-      fs.close();
-
-      cout << "❮ ✔ ❯ Cache is ready" << endl << endl;
-
-
-    }
-
-    /*
-     * Download all fingerprint's data to a variable.
-     * Uses curl to accomplish it.
-     */
-
-    int get_data(){
-      CURL* curl;
-      CURLcode res;
-      curl = curl_easy_init();
-      if (curl == NULL){
-        cerr << "❮ ⚠ ❯ Couldn't get a curl handler!" << endl;
-      }else {
-
-        /* First set the URL that is about to receive our POST. This URL can
-           just as well be a https:// URL if that is what should receive the
-           data. */
-        string body = "laboratory_id=2";
-
-        curl_easy_setopt(curl, CURLOPT_URL,
-            "https://siglaufscar.herokuapp.com/api/fingerprint/");
-        /* Now specify the POST data */
-        curl_easy_setopt(curl, CURLOPT_POSTFIELDS, body.c_str());
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, Digital::write_callback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, this);
-
-        /* Perform the request, res will get the return code */
-        res = curl_easy_perform(curl);
-        /* Check for errors */
-        if (res != CURLE_OK) {
-          cerr << "❮ ⚠ ❯ Could not donwload all fingerprints! (" << curl_easy_strerror(res) << ")" << endl;
-          return 0;
-        }else{
-          cout << "❮ ✔ ❯ Downloaded all fingerpritns!" << endl;
-        }
-
-        this->data = json::parse(this->json_string);
-
-        /* always cleanup */
-        curl_easy_cleanup(curl);
-        return 1;
-      }
-
-    }
-};
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wmissing-noreturn"
 int main() {
